:art: Extension, Component용 파일추가 일부 코드 구조 개선
